Arreglo de BUG a la hora de eliminar POST sin imagen

+++ app/controllers/PostController.php
@@ -300,7 +300,8 @@ class PostController{
     private function delete_Post($pid){
         $publi = new publication();
         $post = json_decode($this->getPost(1, $pid))[0];
-        unlink(addslashes(PUBLIC_DIRECTORY . "images" . DS . $post->Image));
+        if(!is_null($post->Image))
+            unlink(addslashes(PUBLIC_DIRECTORY . "images" . DS . $post->Image));
         $result = $publi->where([["ID_publication", $pid]])->delete();
         if($result){
             return json_encode(["r" => true, "m" => "Se elimino la publicacion satisfactoriamente"], JSON_UNESCAPED_UNICODE);
